animal_dataset.py
from PIL import Image
from torch.utils.data import Dataset


class AnimalDataset(Dataset):
    def __init__(self, data, labels, transform=None):
        super().__init__()
        self.data = data
        self.labels = labels
        self.transform = transform

    def __len__(self):
        return len(self.data)

    def __getitem__(self, index):
        img_name = self.data[index]
        label = self.labels[index]
        image = Image.open(img_name)
        if self.transform is not None:
            image = self.transform(image)
        return image, label, img_name
