pyproject.toml
[tool.poetry]
name = "mlops-2023"
version = "0.1.0"
description = ""
authors = ["VladaTrus <47668051+VladaTrus@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "mlops_2023"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.2"
scikit-learn = "^1.3.2"
pillow = "^10.1.0"
matplotlib = "^3.8.2"
pre-commit = "^3.6.0"
black = "^22.10.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
mlflow = "^2.9.2"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
tqdm = "^4.66.1"
dvc = "^3.36.1"
dvc-gdrive = "^3.0.0"
urllib3 = "1.26.15"
pandas = "^2.1.4"
pyarrow = "^14.0.2"
torchvision = "^0.16.2"

[tool.black]
line-length = 110
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 110
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
